+++ src/main/java/de/kleemann/co2_hsharz/core/transport/TransportMediumService.java
@@ -3,7 +3,7 @@ package de.kleemann.co2_hsharz.core.transport;
 import de.kleemann.co2_hsharz.api.transport.TransportMediumResponseDTO;
 import de.kleemann.co2_hsharz.core.exceptions.CustomEntityNotFoundException;
 import de.kleemann.co2_hsharz.core.exceptions.CustomIllegalArgumentException;
-import de.kleemann.co2_hsharz.persistence.transport.TransportMediumFuel;
+import de.kleemann.co2_hsharz.persistence.fuel.TransportMediumFuel;
 import de.kleemann.co2_hsharz.persistence.transport.TransportMediumName;
 import de.kleemann.co2_hsharz.persistence.transport.TransportMediumPersistenceService;
 import de.kleemann.co2_hsharz.persistence.transport.TransportMediumSize;
