+++ src/main/java/de/kleemann/co2_hsharz/persistence/transport/TransportMediumPersistenceService.java
@@ -5,6 +5,7 @@ import de.kleemann.co2_hsharz.core.exceptions.CustomEntityNotFoundException;
 import de.kleemann.co2_hsharz.core.exceptions.CustomIllegalArgumentException;
 import de.kleemann.co2_hsharz.core.transport.TransportMediumImpl;
 import de.kleemann.co2_hsharz.persistence.attraction.AttractionEntity;
+import de.kleemann.co2_hsharz.persistence.fuel.TransportMediumFuel;
 import jakarta.persistence.EntityExistsException;
 import org.springframework.stereotype.Service;
